Add agent which enables generators to get diffrent pseudo-random numbers
So far each calling of BoundNatural generator it has been returning the same stream of numbers
for the same limit given. For example when we would like to have tuple of two integers with the
same limit the tuple always consists of the same integers.

Now BoundNatural numbers generator returns a stream of diffrent numbers for each calling.
SeedState enables it thanks to keeping a stream of natural numbers generated with start_seed
passed to him with application start. Whenever we call BoundNatural generator it will get different
seed from SeedState.